\name{PropCIs-package}
\Rdversion{0.1-4}
\alias{PropCIs-package}
\docType{package}
\title{
Computes confidence intervals for single proportions, for differences in
proportions, for an odds-ratio and for the relative risk in a 2x2
table. Intervals available for independent samples and matched pairs.
}
\description{
Computes confidence intervals for single proportions, for differences in
proportions, for an odds-ratio and for the relative risk in a 2x2
table. Intervals available for independent samples and matched pairs.
The functions are written by Alan Agresti and students, see http://www.stat.ufl.edu/~aa/cda/cda.html.
}
\details{
\tabular{ll}{
Package: \tab PropCIs\cr
Type: \tab Package\cr
Version: \tab 0.1-4\cr
Date: \tab 2010-02-06\cr
License: \tab GPL=2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ralph Scherer

Maintainer: Ralph Scherer <scherer.ralph@mh-hannover.de>
}
\references{
  Wilson, E.B. 1927. Probable inference, the law of succession, and
  statistical inference. J. Amer. Stat. Assoc. 22:209-212

  Blaker, H. 2000. Confidence curves and improved exact confidence
  intervals for discrete distributions, CANADIAN JOURNAL OF STATISTICS,
  Vol: 28, iss: 4, pg: 783-798

  Olivier, J. and May, W. L. 2006. Weighted confidence interval
  construction for binomial parameters, STATISTICAL METHODS IN MEDICAL
  RESEARCH, Vol: 15, iss:1, pg: 37-46
}
\keyword{ htest }
