\name{exactci}
\Rdversion{0.1-1}
\alias{exactci}
\title{
  Clopper-Pearson exact CI
}
\description{
  Clopper-Pearson exact CI
}
\usage{
exactci(x, n, conflev)
}
\arguments{
  \item{x}{Number of successes}
  \item{n}{Total sample size}
  \item{conflev}{Confidence level}
}
\value{
  \item{ll }{lower confidence bound}
  \item{ul }{upper confidence bound}
}
\keyword{ htest }

