\name{PropCIs-package}
\Rdversion{0.1-5}
\alias{PropCIs-package}
\docType{package}
\title{
Computes confidence intervals for proportions
}
\description{
Computes confidence intervals for single proportions, for differences in
proportions, for an odds-ratio and for the relative risk in a 2x2
table. Intervals available for independent samples and matched pairs.
The functions are written by Alan Agresti and students, see http://www.stat.ufl.edu/~aa/cda/cda.html.
}
\details{
\tabular{ll}{
Package: \tab PropCIs\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2010-02-23\cr
License: \tab GPL=2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ralph Scherer

Maintainer: Ralph Scherer <scherer.ralph@mh-hannover.de>
}
\references{
  Agresti, A., Coull, B. (1998): Approximate is better than exact for interval estimation of binomial proportions. The American Statistician

  Agresti, A., Caffo, B.(2000): Simple and effective confidence
  intervals for proportions and difference of proportions result from
  adding two successes and two failures, The American Statistician

  Agresti, A. 2002. Categorical Data Analysis. Wiley, 2nd Edition.

  A. Agresti and Y. Min, 2004. Improved confidence intervals for
  comparing matched proportions, Statistics in Medicine

  Agresti, A., Gottard, A. (2005): Randomized confidence intervals and
  the mid-P approach, discussion of article by C. Geyer and G. Meeden,
  Statistical Science, vol. 20, pp. 367-371

  D. G. Altman, 1999. Practical statistics for medical research. London,
  Chapman & Hall

  Blaker, H. 2000. Confidence curves and improved exact confidence
  intervals for discrete distributions, CANADIAN JOURNAL OF STATISTICS,
  Vol: 28, iss: 4, pg: 783-798

  Koopman PAR. Confidence limits for the ratio of two binomial
  proportions. Biometrics 1984;40:513-517

  Mee RW. Confidence bounds for the difference between two
  probabilities. Biometrics 1984;40:1175-1176.

  Miettinen OS, Nurminen M. Comparative analysis of two
  rates. Statistics in Medicine 1985;4:213-226.

  Nurminen, 1986. Analysis of trends in proportions with an ordinally
  scaled determinant. Biometrical J. v28. 965-974

  Olivier, J. and May, W. L. 2006. Weighted confidence interval
  construction for binomial parameters, STATISTICAL METHODS IN MEDICAL
  RESEARCH, Vol: 15, iss:1, pg: 37-46

  Wilson, E.B. 1927. Probable inference, the law of succession, and
  statistical inference. J. Amer. Stat. Assoc. 22:209-212
}
\keyword{ htest }
