\name{PropCIs-package}
\Rdversion{0.2-2}
\alias{PropCIs-package}
\docType{package}
\title{
  Confidence intervals for single, paired and independent proportions
}
\description{
Computes confidence intervals for single proportions as well as for
differences in dependent and independent
proportions, the odds-ratio and the relative risk in a 2x2
table. Intervals are available for independent samples and matched pairs.
The functions are partly written by assistants of Alan Agresti, see
website \url{http://www.stat.ufl.edu/~aa/cda/cda.html}.
}
\details{
\tabular{ll}{
Package: \tab PropCIs\cr
Type: \tab Package\cr
Version: \tab 0.2-2\cr
Date: \tab 2013-08-03\cr
License: \tab GPL=2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ralph Scherer

Maintainer: Ralph Scherer <scherer.ralph@mh-hannover.de>
}
\references{
  Agresti, A., Coull, B. (1998) Approximate is better than exact for
  interval estimation of binomial proportions. \emph{The American
    Statistician} 52, 119--126.

  Agresti, A., Caffo, B.(2000) Simple and effective confidence intervals
  for proportions and difference of proportions result from adding two
  successes and two failures. \emph{The American Statistician} 54 (4), 280--288.

  Agresti, A. (2002) Categorical Data Analysis. Wiley, 2nd Edition.

  Agresti, A. and Min, Y. (2005) Simple improved confidence intervals for
  comparing matched proportions \emph{Statistics in Medicine} 24 (5), 729--740.

  Agresti, A., Gottard, A. (2005) Randomized confidence intervals and
  the mid-P approach, discussion of article by C. Geyer and G. Meeden,
  \emph{Statistical Science} 20, 367--371.

  Altman, D. G. (1999) Practical statistics for medical research. London,
  Chapman & Hall.

  Blaker, H. (2000). Confidence curves and improved exact confidence
  intervals for discrete distributions, \emph{Canadian Journal of Statistics}
  28 (4), 783--798.

  Clopper, C. and Pearson, E.S. (1934) The use of cenfidence or fiducial
  limits illustrated in the case of the binomial. \emph{Biometrika} 26,
  404--413.

  Koopman PAR. (1985) Confidence limits for the ratio of two binomial
  proportions. \emph{Biometrics} 40, 513--517.

  Mee, RW. (1984) Confidence bounds for the difference between two
  probabilities. \emph{Biometrics} 40, 1175--1176.

  Miettinen OS, Nurminen M. (1985) Comparative analysis of two
  rates. \emph{Statistics in Medicine} 4, 213--226.

  Nam, J. M. (1995) Confidence limits for the ratio of two binomial
  proportions based on likelihood scores: Non-iterative
  method. \emph{Biom. J.} 37 (3), 375--379.

  Nurminen, M. (1986) Analysis of trends in proportions with an ordinally
  scaled determinant. \emph{Biometrical J.} 28, 965--974.

  Olivier, J. and May, W. L. (2006) Weighted confidence interval
  construction for binomial parameters \emph{Statistical Methods in
  Medical Research} 15 (1), 37--46.

  Tango T. (1998) Equivalence test and confidence interval for the
  difference in proportions for the paired-sample design
  \emph{Statistics in Medicine} 17, 891--908.

  Wilson, E. B. (1927) Probable inference, the law of succession, and
  statistical inference. \emph{J. Amer. Stat. Assoc.} 22, 209--212.
}
\keyword{ htest }
